move keras_model_to_tflite to convert_to_tflite
